Renaming files to fit naming scheme
