+++ workspaces/templates-lib/packages/template-user-management/src/client/getAndPersistToken.ts
@@ -1,5 +1,5 @@
 import { getToken } from './getToken';
-import { getCookieSettings } from './cookieSettings';
+import { getCookieSettings } from './getCookieSettings';
 
 import { setRefreshTokenStorage } from './state';
