+++ workspaces/templates-lib/packages/template-user-management/src/client/performLogout.ts
